+++ web/src/main/java/com/navercorp/pinpoint/web/controller/TransactionController.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2019 NAVER Corp.
+ * Copyright 2025 NAVER Corp.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -18,8 +18,8 @@ package com.navercorp.pinpoint.web.controller;
 
 import com.navercorp.pinpoint.common.hbase.bo.ColumnGetCount;
 import com.navercorp.pinpoint.common.profiler.util.TransactionId;
-import com.navercorp.pinpoint.common.profiler.util.TransactionIdUtils;
 import com.navercorp.pinpoint.common.server.bo.SpanBo;
+import com.navercorp.pinpoint.common.server.util.TransactionIdParser;
 import com.navercorp.pinpoint.common.timeseries.time.Range;
 import com.navercorp.pinpoint.common.timeseries.window.TimeWindow;
 import com.navercorp.pinpoint.web.applicationmap.ApplicationMap;
@@ -144,7 +144,7 @@ public class TransactionController {
                                                      TimeHistogramFormat format) {
         logger.debug("GET /trace params {traceId={}, focusTimestamp={}, agentId={}, spanId={}, format={}}",
                 traceId, focusTimestamp, agentId, spanId, format);
-        final TransactionId transactionId = TransactionIdUtils.parseTransactionId(traceId);
+        final TransactionId transactionId = TransactionIdParser.parseTransactionId(traceId);
         final ColumnGetCount columnGetCount = ColumnGetCount.of(callstackSelectSpansLimit);
 
         final Predicate<SpanBo> spanMatchFilter = SpanFilters.spanFilter(spanId, agentId, focusTimestamp);
@@ -224,7 +224,7 @@ public class TransactionController {
     ) {
         logger.debug("GET /transactionTimelineInfo params {traceId={}, focusTimestamp={}, agentId={}, spanId={}}",
                 traceId, focusTimestamp, agentId, spanId);
-        final TransactionId transactionId = TransactionIdUtils.parseTransactionId(traceId);
+        final TransactionId transactionId = TransactionIdParser.parseTransactionId(traceId);
         final ColumnGetCount columnGetCount = ColumnGetCount.of(callstackSelectSpansLimit);
 
         // select spans
@@ -255,7 +255,7 @@ public class TransactionController {
         logger.debug("GET /traceViewerData params {traceId={}, focusTimestamp={}, agentId={}, spanId={}}",
                 traceIdParam, focusTimestamp, agentId, spanId);
 
-        final TransactionId transactionId = TransactionIdUtils.parseTransactionId(traceIdParam);
+        final TransactionId transactionId = TransactionIdParser.parseTransactionId(traceIdParam);
 
         final ColumnGetCount columnGetCount = ColumnGetCount.of(callstackSelectSpansLimit);
 
@@ -280,7 +280,7 @@ public class TransactionController {
             @RequestParam(value = "useLoadHistogramFormat", required = false, defaultValue = "false")
             boolean useLoadHistogramFormat) {
 
-        final TransactionId transactionId = TransactionIdUtils.parseTransactionId(traceId);
+        final TransactionId transactionId = TransactionIdParser.parseTransactionId(traceId);
         final ColumnGetCount columnGetCount = ColumnGetCount.of(callstackSelectSpansLimit);
         final Predicate<SpanBo> spanMatchFilter = SpanFilters.spanFilter(spanId, agentId, focusTimestamp);
         // select spans
@@ -306,7 +306,7 @@ public class TransactionController {
             boolean useLoadHistogramFormat
     ) {
         TimeHistogramFormat format = TimeHistogramFormat.format(useLoadHistogramFormat);
-        final TransactionId transactionId = TransactionIdUtils.parseTransactionId(traceId);
+        final TransactionId transactionId = TransactionIdParser.parseTransactionId(traceId);
         final ColumnGetCount columnGetCount = ColumnGetCount.of(callstackSelectSpansLimit);
         final Predicate<SpanBo> spanMatchFilter = SpanFilters.spanFilter(spanId, agentId, focusTimestamp);
         // select spans
