+++ web/src/main/java/com/navercorp/pinpoint/web/vo/GetTraceInfoParser.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2020 NAVER Corp.
+ * Copyright 2025 NAVER Corp.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,9 +17,9 @@
 package com.navercorp.pinpoint.web.vo;
 
 import com.navercorp.pinpoint.common.profiler.util.TransactionId;
-import com.navercorp.pinpoint.common.profiler.util.TransactionIdUtils;
-import org.apache.logging.log4j.Logger;
+import com.navercorp.pinpoint.common.server.util.TransactionIdParser;
 import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -52,7 +52,7 @@ public class GetTraceInfoParser {
                 break;
             }
 
-            TransactionId traceId = TransactionIdUtils.parseTransactionId(transactionId);
+            TransactionId traceId = TransactionIdParser.parseTransactionId(transactionId);
             SpanHint spanHint = new SpanHint(Long.parseLong(time), Integer.parseInt(responseTime), applicationName);
 
             final GetTraceInfo getTraceInfo = new GetTraceInfo(traceId, spanHint);
